accept suggestion per gull
